fix(worktree): extract worktree resolver and log pattern errors
- Extract resolve_worktree_info from project.rs (556 lines) into
  project_worktree.rs, bringing project.rs to 480 lines (under 500)
- Replace let _ = with warn! logging for malformed .gitignore and
  .wqmignore patterns (lines 72, 139, 143) so users get feedback on
  invalid exclusion/re-inclusion patterns
